moves reconnect from local to remote actions
